requirements.txt
pandas~=1.5.3
matplotlib~=3.7.1
gradio==3.23.0
prophet~=1.1.2
Pillow~=9.4.0